Processing multiple measurements
git-svn-id: svn://svn.code.sf.net/p/szte-wsn/code/tinyos-2.x/trunk@1013 a4e16f3a-606c-4714-833f-ed604d7aca08

+++ support/cpp/ali/RawToCType/Tracker.cpp
@@ -59,7 +59,7 @@ const string ticks2time(unsigned int t) {
 	sec = t/TICKS_PER_SEC;
 	t   = t%TICKS_PER_SEC;
 
-	milli = t/(TICKS_PER_SEC/1000.0);
+	milli = static_cast<unsigned int> ( t/(TICKS_PER_SEC/1000.0) );
 
 	os << setfill('0') << setw(2) << hour << ":";
 	os << setfill('0') << setw(2) << min  << ":";
